Added functioning config file and ability to load themes.
There is now a file called config.toml that allows you to choose what
rom and theme you want to load. The themes and roms must be put into the
proper folders.
