+++ config.toml
@@ -1,2 +1,2 @@
 rom = "test.ch8"
-theme = "yessir"
+theme = "test"
