--- rom_to_load.txt
@@ -1,1 +1,0 @@
-1-chip8-logo.ch8
